Several changes

+++ build.sh
@@ -5,18 +5,19 @@
 #
 # This uses Spack to install third-party dependencies in the `_spack` directory.
 #
-with_python="OFF"
-python_exe=""
-spack_reinstall=0
-spack_dev=0
 clang=0
 debug="OFF"
+python_exe=""
+spack_dev=0
+spack_env="coekenv"
 spack_home=`pwd`/_spack
+spack_reinstall=0
+with_python="OFF"
 with_valgrind=""
 while [[ $# -gt 0 ]]; do
     case $1 in
         --help)
-                    echo "build.sh [--help] [--clang] [--debug] [--python] [--python-exe <file>] [--spack-dev] [--spack-home <dir>] [--spack-reinstall] [--valgrind]"
+                    echo "build.sh [--help] [--clang] [--debug] [--python] [--python-exe <file>] [--spack-dev] [--spack-env <env>] [--spack-home <dir>] [--spack-reinstall] [--valgrind]"
                     exit 
         ;;
         --clang)
@@ -25,6 +26,11 @@ while [[ $# -gt 0 ]]; do
         ;;
         --debug)
                     debug="ON"
+                    shift
+        ;;
+        --spack-env)
+                    spack_env="$2"
+                    shift
                     shift
         ;;
         --python)
@@ -67,15 +73,21 @@ echo "SPACK_HOME=${SPACK_HOME}"
 if [[ "${spack_reinstall}" -eq 1 ]]; then
     \rm -Rf ${SPACK_HOME}
 fi
-if [[ "$spack_dev" -eq 0 ]]; then
-    \rm -Rf _spack_tpls
-    git clone https://github.com/or-fusion/or-fusion-spack-repo.git _spack_tpls
-else
-    git clone git@github.com:or-fusion/or-fusion-spack-repo.git _spack_tpls
+if test -d _spack_tpls; then
     cd _spack_tpls
-    git checkout dev
     git pull
     cd ..
+else
+    if [[ "$spack_dev" -eq 0 ]]; then
+        \rm -Rf _spack_tpls
+        git clone https://github.com/or-fusion/or-fusion-spack-repo.git _spack_tpls
+    else
+        git clone git@github.com:or-fusion/or-fusion-spack-repo.git _spack_tpls
+        cd _spack_tpls
+        git checkout dev
+        git pull
+        cd ..
+    fi
 fi
 \rm -Rf _build
 #
@@ -106,11 +118,14 @@ else
     echo ""
     git clone https://github.com/spack/spack.git ${SPACK_HOME}
     . ${SPACK_HOME}/share/spack/setup-env.sh
+    echo "Adding _spack_tpls"
+    spack repo remove _spack_tpls/repo | true
     spack repo add _spack_tpls/repo
     spack repo list
-    spack env create coekenv
-    spack env activate coekenv
-    spack add asl cppad fmt rapidjson catch2 highs $with_valgrind
+    spack env create $spack_env
+    spack env activate $spack_env
+    spack compiler find
+    spack add asl fmt rapidjson catch2 highs $with_valgrind
     spack install
     spack env deactivate
     spack repo remove _spack_tpls/repo
@@ -126,5 +141,7 @@ echo "Building Coek"
 echo ""
 mkdir _build
 cd _build
-cmake -DCMAKE_PREFIX_PATH=${SPACK_HOME}/var/spack/environments/coekenv/.spack-env/view -Dwith_debug=${debug} -Dwith_python=${with_python} $python_exe -Dwith_gurobi=$with_gurobi -Dwith_highs=ON -Dwith_cppad=ON -Dwith_fmtlib=ON -Dwith_rapidjson=ON -Dwith_catch2=ON -Dwith_tests=ON -Dwith_asl=ON -Dwith_openmp=OFF ..
+cmake -DCMAKE_PREFIX_PATH=${SPACK_HOME}/var/spack/environments/${spack_env}/.spack-env/view -Dwith_debug=${debug} -Dwith_python=${with_python} $python_exe -Dwith_gurobi=$with_gurobi -Dwith_highs=ON -Dwith_cppad=OFF -Dwith_fmtlib=ON -Dwith_rapidjson=ON -Dwith_catch2=ON -Dwith_tests=ON -Dwith_asl=ON -Dwith_openmp=OFF ..
 make -j20
+make install
+
